ml-pipeline/requirements.txt
# Core ML Libraries
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3

# Data Processing
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
joblib==1.3.2

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0

# Model Monitoring
mlflow==2.8.1 